Backend/requirements.txt
google-genai==1.59.0
python-dotenv==1.2.1
flask==3.1.2
flask-cors==6.0.2
